make it work with GCC 4.4, fix some signedness warnings

+++ dumux/boxmodels/3p3c/3p3cmodel.hh
@@ -228,7 +228,7 @@ public:
         ParentType::updateFailed();
 
         auto &sol = this->solution(/*timeIdx=*/0);
-        for (int i = 0; i < sol.size(); ++i)
+        for (size_t i = 0; i < sol.size(); ++i)
             sol[i].setSwitched(false);
         numSwitched_ = 0;
     }
